LayoutTests/fast/forms/datalist/datalist-textinput-dynamically-add-options-on-keydown-expected.txt


PASS input.value became 'A'
PASS isShowingDataListSuggestions is true
PASS successfullyParsed is true

TEST COMPLETE
